front: drop unused code from last itinerary interfaces
