--- front/src/applications/operationalStudies/views/Scenario/components/ManageTrainSchedule/ManageTrainScheduleMap/index.ts
@@ -1,3 +1,0 @@
-import ManageTrainScheduleMap from './ManageTrainScheduleMap';
-
-export default ManageTrainScheduleMap;
