--- front/src/applications/operationalStudies/views/Scenario/components/ManageTrainSchedule/PowerRestrictionsSelector/helpers/utils.ts
@@ -1,55 +1,0 @@
-import type { ManageTrainSchedulePathProperties } from 'applications/operationalStudies/types';
-import type { TrackSection } from 'common/api/osrdEditoastApi';
-import type { IntervalItem } from 'common/IntervalsEditor/types';
-import type { PathStep } from 'reducers/osrdconf/types';
-import { mToMm } from 'utils/physics';
-
-import createPathStep from './createPathStep';
-
-/** Get pathStep located at the given position on path (in meters), if it exists. */
-export const getPathStep = (pathSteps: PathStep[], positionInM: number) =>
-  pathSteps.find((step) => step.positionOnPath === mToMm(positionInM));
-
-export const getOrCreatePathStepAtPosition = (
-  positionOnPathInM: number,
-  pathSteps: PathStep[],
-  tracksLengthCumulativeSums: number[],
-  pathProperties: ManageTrainSchedulePathProperties,
-  tracksById: Record<string, TrackSection>
-) => {
-  const pathStep = getPathStep(pathSteps, positionOnPathInM);
-  if (pathStep) {
-    return pathStep;
-  }
-  return createPathStep(
-    positionOnPathInM,
-    tracksLengthCumulativeSums,
-    pathProperties,
-    pathSteps,
-    tracksById
-  );
-};
-
-export const extractPathStepsFromRange = (
-  range: IntervalItem,
-  pathSteps: PathStep[],
-  tracksLengthCumulativeSums: number[],
-  pathProperties: ManageTrainSchedulePathProperties,
-  tracksById: Record<string, TrackSection>
-) => {
-  const from = getOrCreatePathStepAtPosition(
-    range.begin,
-    pathSteps,
-    tracksLengthCumulativeSums,
-    pathProperties,
-    tracksById
-  );
-  const to = getOrCreatePathStepAtPosition(
-    range.end,
-    pathSteps,
-    tracksLengthCumulativeSums,
-    pathProperties,
-    tracksById
-  );
-  return { from, to };
-};
