--- front/src/applications/operationalStudies/views/Scenario/consts.ts
@@ -1,6 +1,0 @@
-export const MANAGE_TRAIN_SCHEDULE_TYPES = Object.freeze({
-  none: 'NONE',
-  add: 'ADD',
-  edit: 'EDIT',
-  itinerary: 'ITINERARY',
-});
