--- front/src/common/IntervalsEditor/index.ts
@@ -1,3 +1,0 @@
-import IntervalsEditor from './IntervalsEditor';
-
-export default IntervalsEditor;
