--- front/src/modules/pathfinding/types.ts
@@ -1,6 +1,0 @@
-export type PathfindingState = {
-  isMissingParam: boolean;
-  isRunning: boolean;
-  isDone: boolean;
-  error?: string;
-};
