+++ front/src/modules/timesStops/consts.ts
@@ -1,7 +1,5 @@
 import { Duration } from 'utils/duration';
 
-export const marginRegExValidation = /^(\d+(\.\d+)?)(%|min\/100km)$/;
-
 export const MarginUnit = {
   percent: '%',
   minPer100km: 'min/100km',
