+++ front/src/modules/timesStops/helpers/powerRestrictionIncompatibility.ts
@@ -1,7 +1,7 @@
 import type { PathPropertiesFormatted } from 'applications/operationalStudies/types';
-import { getPowerRestrictionsWarnings } from 'applications/operationalStudies/views/Scenario/components/ManageTrainSchedule/PowerRestrictionsSelector/helpers/powerRestrictionWarnings';
 import type { PowerRestrictionItem, RollingStock } from 'common/api/osrdEditoastApi';
 import { NO_POWER_RESTRICTION } from 'modules/powerRestriction/consts';
+import { getPowerRestrictionsWarnings } from 'modules/timesStops/helpers/powerRestrictionWarnings';
 import type { Train } from 'reducers/osrdconf/types';
 
 import type { TimesStopsRowNew } from '../types';
