+++ front/src/modules/timesStops/helpers/utils.ts
@@ -1,7 +1,5 @@
-import { keyColumn, createTextColumn } from '@sdziadkowiec/react-datasheet-grid';
-import dayjs from 'dayjs';
 import type { TFunction } from 'i18next';
-import { round, isEqual, isNil } from 'lodash';
+import { isNil } from 'lodash';
 
 import { getInvalidStepLabel } from 'applications/operationalStudies/utils';
 import type {
@@ -9,21 +7,7 @@ import type {
   PathItemLocation,
   ReceptionSignal,
 } from 'common/api/osrdEditoastApi';
-import type { TimeString } from 'common/types';
-import type { SuggestedOP } from 'modules/trainSchedule/types';
-import type { PathStep } from 'reducers/osrdconf/types';
 import { Duration, type StartTime } from 'utils/duration';
-import { msToS } from 'utils/physics';
-import { NO_BREAK_SPACE } from 'utils/strings';
-import {
-  durationInSeconds,
-  SECONDS_IN_A_DAY,
-  secToHoursString,
-  time2sec,
-} from 'utils/timeManipulation';
-
-import { marginRegExValidation, MarginUnit } from '../consts';
-import { TableType, type TimeExtraDays, type TimesStopsInputRow } from '../types';
 
 export const truncateStartTimeToSecond = (date: StartTime): StartTime => {
   if (date instanceof Date) {
@@ -55,324 +39,6 @@ export function receptionSignalToSignalBooleans(receptionSignal?: ReceptionSigna
     return { shortSlipDistance: true, onStopSignal: true };
   }
   return { shortSlipDistance: false, onStopSignal: false };
-}
-
-export function calculateStepTimeAndDays(
-  startTime?: Date | null,
-  duration?: Duration | null
-): TimeExtraDays | undefined {
-  if (!startTime || !duration) {
-    return undefined;
-  }
-
-  const start = dayjs(startTime);
-  const dur = dayjs.duration(duration.ms);
-
-  const waypointArrivalTime = start.add(dur);
-  const daySinceDeparture = waypointArrivalTime.diff(start, 'day');
-  const time: TimeString = waypointArrivalTime.format('HH:mm:ss');
-
-  return {
-    time,
-    daySinceDeparture,
-  };
-}
-
-export const formatSuggestedViasToRowVias = (
-  operationalPoints: SuggestedOP[],
-  pathSteps: PathStep[],
-  t: TFunction<'translation', undefined>,
-  startTime?: Date,
-  tableType?: TableType
-): TimesStopsInputRow[] => {
-  const formattedOps = [...operationalPoints];
-
-  // If the origin is in the ops and isn't the first operational point, we need
-  // to move it to the first position
-  const origin = pathSteps[0];
-  const originIndexInOps = origin
-    ? operationalPoints.findIndex((op) => origin.id === op.pathStepId)
-    : -1;
-  if (originIndexInOps !== -1) {
-    [formattedOps[0], formattedOps[originIndexInOps]] = [
-      formattedOps[originIndexInOps],
-      formattedOps[0],
-    ];
-  }
-
-  // Ditto: destination should be last
-  const dest = pathSteps[pathSteps.length - 1];
-  const destIndexInOps = dest ? operationalPoints.findIndex((op) => dest.id === op.pathStepId) : -1;
-  if (destIndexInOps !== -1) {
-    const lastOpIndex = formattedOps.length - 1;
-    [formattedOps[lastOpIndex], formattedOps[destIndexInOps]] = [
-      formattedOps[destIndexInOps],
-      formattedOps[lastOpIndex],
-    ];
-  }
-
-  return formattedOps.map((op, i) => {
-    const pathStep = pathSteps.find((step) => step.id === op.pathStepId);
-    const name = pathStep?.name || op.name;
-    const objectToUse = tableType === TableType.Input ? pathStep : op;
-
-    const { arrival, receptionSignal, stopFor, theoreticalMargin } = objectToUse || {};
-
-    const stopForSeconds = stopFor ? stopFor.total('second') : undefined;
-
-    const isMarginValid = theoreticalMargin ? marginRegExValidation.test(theoreticalMargin) : true;
-    const arrivalDuration = i === 0 ? Duration.zero : arrival;
-    const arrivalInSeconds = arrivalDuration ? msToS(arrivalDuration.ms) : null;
-
-    const formattedArrival = calculateStepTimeAndDays(startTime, arrivalDuration);
-
-    const departureTime =
-      stopForSeconds !== undefined && arrivalInSeconds
-        ? secToHoursString(arrivalInSeconds + stopForSeconds)
-        : undefined;
-    const formattedDeparture: TimeExtraDays | undefined = departureTime
-      ? { time: departureTime }
-      : undefined;
-    const { receptionSignal: _opReceptionSignal, ...filteredOp } = op;
-    const { shortSlipDistance, onStopSignal } = receptionSignalToSignalBooleans(receptionSignal);
-    return {
-      ...filteredOp,
-      isMarginValid,
-      arrival: formattedArrival,
-      departure: formattedDeparture,
-      onStopSignal,
-      name: name || t('timeStopTable.waypoint', { id: filteredOp.pathStepId }),
-      shortSlipDistance,
-      stopFor,
-      theoreticalMargin,
-    };
-  });
-};
-
-const getDigits = (unit: string | undefined) =>
-  unit === MarginUnit.second || unit === MarginUnit.percent ? 0 : 1;
-
-export function formatDigitsAndUnit(fullValue: string | number | undefined, unit?: string) {
-  if (fullValue === undefined) {
-    return '';
-  }
-  if (typeof fullValue === 'number') {
-    return `${round(Number(fullValue), getDigits(unit))}${NO_BREAK_SPACE}${unit}`;
-  }
-  const splitValue = fullValue.match(marginRegExValidation);
-  if (!splitValue) {
-    return '';
-  }
-  const extractedValue = Number(splitValue[1]);
-  const extractedUnit = splitValue[3];
-  const digits = getDigits(extractedUnit);
-  return `${round(extractedValue, digits)}${NO_BREAK_SPACE}${extractedUnit}`;
-}
-
-export function disabledTextColumn(
-  key: string,
-  title: string,
-  options?: Parameters<typeof createTextColumn>[0]
-) {
-  return {
-    ...keyColumn(key, createTextColumn(options)),
-    title,
-    disabled: true,
-  };
-}
-
-/**
- * Synchronizes arrival, departure and stop times.
- * updates onStopSignal
- * updates isMarginValid and theoreticalMargin
- */
-export function updateRowTimesAndMargin(
-  rowData: TimesStopsInputRow,
-  previousRowData: TimesStopsInputRow,
-  op: { fromRowIndex: number },
-  allWaypointsLength: number
-): TimesStopsInputRow {
-  const newRowData = { ...rowData };
-  const {
-    stopFor: previousStopFor,
-    departure: previousDeparture,
-    arrival: previousArrival,
-  } = previousRowData;
-  const arrivalChanged = !isEqual(newRowData.arrival, previousRowData.arrival);
-  const departureChanged = !isEqual(newRowData.departure, previousRowData.departure);
-  const stopDurationChanged = !isEqual(newRowData.stopFor, previousRowData.stopFor);
-
-  if (previousArrival?.time && previousDeparture?.time) {
-    if (!newRowData.arrival?.time || !newRowData.departure?.time) {
-      newRowData.departure = undefined;
-      newRowData.arrival = undefined;
-    } else {
-      if (arrivalChanged || stopDurationChanged) {
-        if (stopDurationChanged && !newRowData.stopFor) {
-          newRowData.departure = undefined;
-        } else {
-          newRowData.departure = {
-            time: secToHoursString(
-              time2sec(newRowData.arrival.time) + (newRowData.stopFor?.total('second') ?? 0)
-            ),
-          };
-        }
-      }
-      if (departureChanged) {
-        if (newRowData.departure?.time) {
-          newRowData.stopFor = new Duration({
-            seconds: durationInSeconds(
-              time2sec(newRowData.arrival.time),
-              time2sec(newRowData.departure.time)
-            ),
-          });
-        }
-      }
-    }
-  } else if (previousArrival?.time) {
-    if (newRowData.arrival?.time && (arrivalChanged || stopDurationChanged)) {
-      if ((newRowData.stopFor?.total('second') ?? 0) > 0) {
-        newRowData.departure = {
-          time: secToHoursString(
-            time2sec(newRowData.arrival.time) + (newRowData.stopFor?.total('second') ?? 0)
-          ),
-        };
-      }
-    }
-    if (newRowData.departure?.time && newRowData.arrival?.time) {
-      if (newRowData.arrival?.time) {
-        newRowData.stopFor = new Duration({
-          seconds: durationInSeconds(
-            time2sec(newRowData.arrival.time),
-            time2sec(newRowData.departure.time)
-          ),
-        });
-      }
-    }
-  }
-
-  if (previousStopFor && !previousDeparture?.time && newRowData.departure?.time) {
-    newRowData.arrival = {
-      time: secToHoursString(
-        time2sec(newRowData.departure.time) - (newRowData.stopFor?.total('second') ?? 0)
-      ),
-    };
-  }
-
-  if (
-    !newRowData.stopFor &&
-    newRowData.onStopSignal &&
-    op.fromRowIndex !== allWaypointsLength - 1
-  ) {
-    newRowData.onStopSignal = false;
-  }
-  newRowData.isMarginValid = !(
-    newRowData.theoreticalMargin && !marginRegExValidation.test(newRowData.theoreticalMargin)
-  );
-  if (newRowData.isMarginValid && op.fromRowIndex === 0) {
-    newRowData.arrival = undefined;
-    // As we put 0% by default for origin's margin, if the user removes a margin without
-    // replacing it to 0% (undefined), we change it to 0%
-    if (!newRowData.theoreticalMargin) {
-      newRowData.theoreticalMargin = '0%';
-    }
-  }
-  return newRowData;
-}
-
-/**
- * This function is called before comparing rows to prevent a change from undefined to null (or the reverse)
- * from being treated as an actual update of a row (otherwise changes would occur on deletion of an undefined field)
- */
-export function normalizeNullablesInRow(row: TimesStopsInputRow): TimesStopsInputRow {
-  const normalizedRow = { ...row };
-  if (normalizedRow.stopFor === null) {
-    normalizedRow.stopFor = undefined;
-  }
-  if (normalizedRow.theoreticalMargin === null) {
-    normalizedRow.theoreticalMargin = undefined;
-  }
-  return normalizedRow;
-}
-
-/**
- * This function goes through the whole array of path waypoints
- * and updates the number of days since departure.
- */
-export function updateDaySinceDeparture(
-  pathWaypointRows: TimesStopsInputRow[],
-  { keepFirstIndexArrival = false } = {}
-): TimesStopsInputRow[] {
-  let currentDaySinceDeparture = 0;
-  let previousTime = Number.NEGATIVE_INFINITY;
-
-  return pathWaypointRows.map((pathWaypoint, index) => {
-    const { arrival, stopFor } = pathWaypoint;
-    const arrivalInSeconds = arrival?.time ? time2sec(arrival.time) : null;
-    let formattedArrival: TimeExtraDays | undefined;
-
-    if (arrivalInSeconds !== null) {
-      const isMidnight = arrival?.time === '00:00:00';
-
-      if ((arrivalInSeconds < previousTime || isMidnight) && !(isMidnight && index === 0)) {
-        currentDaySinceDeparture += 1;
-        formattedArrival = {
-          time: arrival!.time,
-          daySinceDeparture: currentDaySinceDeparture,
-          dayDisplayed: true,
-        };
-      } else {
-        formattedArrival = {
-          time: arrival!.time,
-          daySinceDeparture: currentDaySinceDeparture,
-        };
-      }
-      previousTime = isMidnight ? 0 : arrivalInSeconds;
-    }
-
-    let formattedDeparture: TimeExtraDays | undefined;
-
-    if (stopFor && arrivalInSeconds !== null) {
-      const departureInSeconds = (arrivalInSeconds + stopFor.total('second')) % SECONDS_IN_A_DAY;
-      const isAfterMidnight = departureInSeconds < previousTime;
-      const isDepartureMidnight = departureInSeconds === 0;
-
-      if (isAfterMidnight || isDepartureMidnight) {
-        currentDaySinceDeparture += 1;
-        formattedDeparture = {
-          time: secToHoursString(departureInSeconds),
-          daySinceDeparture: currentDaySinceDeparture,
-          dayDisplayed: true,
-        };
-      } else {
-        formattedDeparture = {
-          time: secToHoursString(departureInSeconds),
-          daySinceDeparture: currentDaySinceDeparture,
-        };
-      }
-      previousTime = departureInSeconds;
-    }
-    return {
-      ...pathWaypoint,
-      arrival: keepFirstIndexArrival || index > 0 ? formattedArrival : undefined,
-      departure: formattedDeparture,
-    };
-  });
-}
-
-export function durationSinceStartTime(
-  startTime?: Date,
-  stepTimeDays?: TimeExtraDays
-): Duration | null {
-  if (!startTime || !stepTimeDays?.time || stepTimeDays?.daySinceDeparture === undefined) {
-    return null;
-  }
-  const start = dayjs(startTime);
-  const step = dayjs(`${start.format('YYYY-MM-DD')}T${stepTimeDays.time}`).add(
-    stepTimeDays.daySinceDeparture,
-    'day'
-  );
-  return Duration.subtractDate(step.toDate(), startTime);
 }
 
 /** Convert onStopSignal boolean to receptionSignal enum */
